% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantities.R
\name{adjustQscale}
\alias{adjustQscale}
\title{Adjust a quantity scale to a value and/or prefix and unit constraints}
\usage{
adjustQscale(x, to = NULL, mode = "fraction", prefix = NULL,
  unit = NULL)
}
\arguments{
\item{x}{a quantity}

\item{to}{a numerical 'target' value}

\item{mode}{character string which may have the following values 'max', if 'to' is the maximum value,
'min' if 'to' is the minimum value, 'round, if 'x' should be 'rounded' to 'to', and 'fraction'
if the fraction 'x'/'to' should be minimized.}

\item{prefix}{NULL or a characxter vector with the prefixes allowed. The vector may coantain
two special values '=' and'1000s'. '=' denotes the prefix of x, '1000s' all prefixes except
'hecta', 'deca', 'deci', and 'centi'. If NULL all prefixes are allowed.}

\item{unit}{NULL or a character vector with the units allowed. The vector may contain also a '=',
denoting the unit of x. If NULL all units are allowed.}
}
\description{
Adjust a quantity scale to a value and/or prefix and unit constraints
}
\details{
The mode, prefix and unit values may be abbreviated. \cr
The only quantity affected is the first (non-prefix) quantity in the quantity defintion.
}
