% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{as.data.frame.quantities}
\alias{as.data.frame.quantities}
\title{Coerce to a Data Frame}
\usage{
\method{as.data.frame}{quantities}(x, ...)
}
\arguments{
\item{x}{any \R object.}

\item{...}{additional arguments to be passed to or from methods.}
}
\description{
S3 method for \code{quantities} objects (see \code{\link{as.data.frame}}).
}
\examples{
x <- set_quantities(1:3, m/s, 0.1)
y <- set_quantities(4:6, m/s, 0.2)
(z <- cbind(x, y))
as.data.frame(z)

}
