% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{as.list.quantities}
\alias{as.list.quantities}
\title{Coerce to a List}
\usage{
\method{as.list}{quantities}(x, ...)
}
\arguments{
\item{x}{object to be coerced or tested.}

\item{...}{objects, possibly named.}
}
\description{
S3 method for \code{quantities} objects (see \code{\link{as.list}}).
}
\examples{
x <- set_quantities(1:3, m/s, 0.1)
as.list(x)

}
