% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{as.matrix.quantities}
\alias{as.matrix.quantities}
\title{Coerce to a Matrix}
\usage{
\method{as.matrix}{quantities}(x, ...)
}
\arguments{
\item{x}{an \R object.}

\item{...}{additional arguments to be passed to or from methods.}
}
\description{
S3 method for \code{quantities} objects (see \code{\link{as.matrix}}).
}
\examples{
as.matrix(set_quantities(1:3, m/s, 0.1))

}
