% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop.R
\name{drop_quantities}
\alias{drop_quantities}
\alias{drop_units.quantities}
\alias{drop_errors.quantities}
\title{Drop Units and Errors}
\usage{
drop_quantities(x)

\method{drop_units}{quantities}(x)

\method{drop_errors}{quantities}(x)
}
\arguments{
\item{x}{a \code{quantities} object.}
}
\value{
the numeric without any \code{units} or \code{errors} attributes,
while preserving other attributes like dimensions or other classes.
}
\description{
Drop Units and Errors
}
\details{
\code{drop_quantities} is equivalent to \code{quantities(x) <- NULL}
or \code{set_quantities(x, NULL, NULL)}. \code{drop_units} is equivalent to
\code{units(x) <- NULL} or \code{set_units(x, NULL)}. \code{drop_errors} is
equivalent to \code{errors(x) <- NULL} or \code{set_errors(x, NULL)}.
}
