% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numeric.R
\name{numericAndInteger}
\alias{numericAndInteger}
\alias{is.wholeNumber}
\alias{asNumeric}
\alias{asInteger}
\title{coerce a value to a numeric or an integer}
\usage{
is.wholeNumber(x)

asNumeric(x)

asInteger(x)
}
\arguments{
\item{x}{an R vector value to be coerced}
}
\value{
the numeric / integer value or NA if x could not be converted
}
\description{
coerce a value to a numeric or an integer
}
\details{
asNumeric is the same as as.Numeric with suppressed warnings. \cr
asInteger does not perform any rounding. A NA is returned in stead. \cr
is.wholeNUmber returns TRUE is its argument can be converted to an integer without rounding
}
