% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{parse_quantities}
\alias{parse_quantities}
\alias{parse_units}
\alias{parse_errors}
\title{Parse Units and Errors}
\usage{
parse_quantities(x)

parse_units(x)

parse_errors(x)
}
\arguments{
\item{x}{a character vector to parse.}
}
\value{
A \code{quantities}, \code{units} or \code{errors} object respectively.
}
\description{
Functions to parse character vectors into quantities.
}
\details{
Each \code{parse_*()} function returns an object of the corresponding
type, no matter what it is found. This means that, for \code{parse_units}, if
errors are found, they are dropped with a warning. Similarly for
\code{parse_errors}, if units are found, they are dropped with a warning.
On the other hand, \code{parse_quantities} always returns a valid
\code{quantities} object, even if no errors or units are found (then, zero
error and dimensionless units are applied).
}
\examples{
parse_quantities("(1.6021766208 +/- .0000000098) e-19 C")
parse_quantities("1.6021766208(98) e-19 C")
parse_units("1.6021766208 e-19 C")
parse_errors("1.6021766208(98) e-19")

}
