% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{procesEnumertedParameter}
\alias{procesEnumertedParameter}
\title{process a possible abbreviated enumetered type parameter}
\usage{
procesEnumertedParameter(x, type)
}
\arguments{
\item{x}{a character string with the actual paremeter}

\item{type}{a character vector with the set of possible values}
}
\value{
the unabreviated value or as.character(NA) in case of an error
}
\description{
process a possible abbreviated enumetered type parameter
}
\keyword{internal}
