% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantityDefinitions.R
\name{quanityDefintions}
\alias{quanityDefintions}
\alias{addQdefinition}
\alias{printQnames}
\alias{printQunits}
\title{Qunatities definitions}
\usage{
addQdefinition(qName, units, default)

printQnames()

printQunits(qName)
}
\arguments{
\item{qName}{cheracter string, the name of quantity}

\item{units}{a list with unit definitions}

\item{default}{the name of the defeault unit.}
}
\description{
Qunatities definitions
}
\details{
addQdefinition adds /replaces a guantity efintion  \cr
printQnames prints the namesof quantities defined \cr
printQunits prints th equantity units and other quantity details
}
