% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantities.R
\docType{package}
\name{quantities-package}
\alias{quantities-package}
\title{\pkg{quantities}: Quantity Calculus for R Vectors}
\description{
Support for painless automatic units and uncertainty propagation in numerical
operations. Both \pkg{units} and \pkg{errors} are integrated into a complete
quantity calculus system within the R language. R vectors, matrices and arrays
automatically propagate those attributes when you operate with \code{quantities}
objects.
}
\references{
Edzer Pebesma, Thomas Mailund and James Hiebert (2016).
Measurement Units in \R. \emph{The R Journal}, 8(2), 486--494.
\doi{10.32614/RJ-2016-061}

Iñaki Ucar, Edzer Pebesma and Arturo Azcorra (2018).
Measurement Errors in \R. \emph{The R Journal}, 10(2), 549--557.
\doi{10.32614/RJ-2018-075}
}
\author{
Iñaki Ucar
}
