% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numeric.R
\name{roundDecimals}
\alias{roundDecimals}
\title{rounding the decimals to a given number of significant numeric digits}
\usage{
roundDecimals(x, signif = 6)
}
\arguments{
\item{x}{a numeric vector to be rounded}

\item{signif}{the number of significant numeric digits. default = 6}
}
\value{
the rounded number
}
\description{
rounding the decimals to a given number of significant numeric digits
}
\keyword{internal}
