% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{t.quantities}
\alias{t.quantities}
\title{Matrix Transpose}
\usage{
\method{t}{quantities}(x)
}
\arguments{
\item{x}{a matrix or data frame, typically.}
}
\description{
S3 method for \code{quantities} objects (see \code{\link{t}}).
}
\examples{
a <- matrix(1:30, 5, 6)
quantities(a) <- list("m/s", 1:30)
t(a)

}
