% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{tibble}
\alias{tibble}
\alias{type_sum.quantities}
\alias{pillar_shaft.quantities}
\title{Methods for Tidy \code{tibble} Printing}
\usage{
type_sum.quantities(x)

pillar_shaft.quantities(x, ...)
}
\arguments{
\item{x}{object of class quantities.}

\item{...}{see \link[pillar]{pillar_shaft}.}
}
\description{
S3 methods for \code{quantities} objects.
}
