% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{type_sum.quantities}
\alias{type_sum.quantities}
\title{\code{type_sum} for Tidy \code{tibble} Printing}
\usage{
type_sum.quantities(x, ...)
}
\arguments{
\item{x}{object of class quantities}

\item{...}{ignored}
}
\description{
S3 method for \code{quantities} objects.
}
