% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errors.R
\name{errors}
\alias{errors}
\alias{errors<-.quantities}
\alias{errors}
\alias{errors<-.units}
\alias{errors}
\alias{set_errors.quantities}
\alias{errors}
\alias{set_errors.units}
\alias{errors}
\alias{errors_max.quantities}
\alias{errors}
\alias{errors_min.quantities}
\title{Set Measurement Errors on a Numeric Vector}
\usage{
\method{errors}{quantities}(x) <- value

\method{errors}{units}(x) <- value

\method{set_errors}{quantities}(x, value = 0)

\method{set_errors}{units}(x, value = 0)

\method{errors_max}{quantities}(x)

\method{errors_min}{quantities}(x)
}
\arguments{
\item{x}{a numeric object, or object of class \code{errors}.}

\item{value}{a numeric vector of length 1 or the same length as \code{x}.}
}
\description{
Set/retrieve measurement errors to/from numeric vectors (extensions to the
\code{errors} package for \code{quantities} and \code{units} objects).
}
\seealso{
\code{\link[errors]{errors}}.
}
