% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantities.R
\docType{package}
\name{quantities-package}
\alias{quantities-package}
\title{\pkg{quantities}: Quantity Calculus for R Vectors}
\description{
Support for painless automatic units and uncertainty propagation in numerical
operations. Both \pkg{units} and \pkg{errors} are integrated into a complete
quantity calculus system within the R language. R vectors, matrices and arrays
automatically propagate those attributes when you operate with \code{quantities}
objects.
}
\author{
Iñaki Ucar
}
