% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantities.R
\name{units}
\alias{units}
\alias{units<-.quantities}
\alias{units}
\alias{units<-.errors}
\alias{units}
\alias{set_units.quantities}
\alias{units}
\alias{set_units.errors}
\title{Set Measurement Units on a Numeric Vector}
\usage{
\method{units}{quantities}(x) <- value

\method{units}{errors}(x) <- value

\method{set_units}{quantities}(x, value, ...)

\method{set_units}{errors}(x, value, ...)
}
\arguments{
\item{x}{numeric vector, or object of class \code{units}}

\item{value}{object of class \code{units} or \code{symbolic_units}, or in the case of \code{set_units} expression with symbols that can be resolved in \link{ud_units} (see examples).}

\item{...}{passed on to \code{as_units}}
}
\description{
Set/retrieve measurement units to/from numeric vectors and convert units
(extensions to the \code{units} package for \code{quantities} and
\code{errors} objects).
}
\seealso{
\code{\link[units]{units}}, \code{\link[units]{set_units}}.
}
